requirements.txt
Flask==1.1.1
Flask-Compress==1.5.0
flask-restplus==0.13.0
behave==1.2.6
PyHamcrest==1.9.0
allure-behave==2.8.5
matplotlib==3.1.3
mpmath==1.1.0
numpy==1.18.1
pandas==1.0.1
py-ard==0.6.1
requests==2.22.0
seaborn==0.10.0
toolz==0.11.1
sigfig==1.1.8
Werkzeug==0.16.1
